# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BrewAsset POS & WMS API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface LoginRequest {
  username: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  rep: "rep",
  warehouse: "warehouse",
  marketing: "marketing",
} as const;

export interface User {
  id: number;
  username: string;
  role: UserRole;
  displayName: string;
}

export interface LoginResponse {
  user: User;
}

export interface InventoryItem {
  id: number;
  itemType: string;
  brand: string;
  quantity: number;
  catalogItemId: number;
  brandId: number;
}

export interface ReceiveStockRequest {
  itemType: string;
  brand: string;
  quantity: number;
}

export interface ReceiveHistoryItem {
  id: number;
  itemType: string;
  brand: string;
  quantity: number;
  note?: string | null;
  createdAt: string;
}

export type PosRequestStatus =
  (typeof PosRequestStatus)[keyof typeof PosRequestStatus];

export const PosRequestStatus = {
  Pending: "Pending",
  Fulfilled: "Fulfilled",
  Cancelled: "Cancelled",
} as const;

export interface PosRequest {
  id: number;
  repUsername: string;
  customer: string;
  itemType: string;
  brand: string;
  quantity: number;
  status: PosRequestStatus;
  createdAt: string;
  fulfilledAt?: string | null;
}

export interface CreateRequestBody {
  customer: string;
  itemType: string;
  brand: string;
  quantity: number;
}

export type EventStatus = (typeof EventStatus)[keyof typeof EventStatus];

export const EventStatus = {
  pending: "pending",
  approved: "approved",
  cancelled: "cancelled",
} as const;

export interface Event {
  id: number;
  title: string;
  account: string;
  brand?: string | null;
  date: string;
  startTime: string;
  endTime: string;
  notes?: string | null;
  posItems: string[];
  staffRequested: number;
  repUsername: string;
  status: EventStatus;
  staffAssigned: string[];
  posApproved: boolean;
  posRequested: boolean;
  createdAt: string;
  approvedAt?: string | null;
  approvedBy?: string | null;
}

export interface CreateEventBody {
  title: string;
  account: string;
  brand?: string | null;
  date: string;
  startTime: string;
  endTime: string;
  notes?: string | null;
  posItems: string[];
  staffRequested: number;
}

export interface ToggleStaffBody {
  staffName: string;
}

export interface Transfer {
  id: number;
  repUsername: string;
  itemType: string;
  brand: string;
  quantity: number;
  fromAccount: string;
  toAccount: string;
  isReturn: boolean;
  createdAt: string;
}

export interface CreateTransferBody {
  itemType: string;
  brand: string;
  quantity: number;
  fromAccount: string;
  toAccount: string;
}

export interface PromoStaff {
  id: number;
  name: string;
  phone?: string | null;
  email?: string | null;
  notes?: string | null;
}

export interface CreatePromoStaffBody {
  name: string;
  phone?: string | null;
  email?: string | null;
  notes?: string | null;
}

export interface Brand {
  id: number;
  name: string;
  catalogItemId: number;
  isDefault: boolean;
}

export interface CatalogItem {
  id: number;
  name: string;
  icon?: string | null;
  color?: string | null;
  isDefault: boolean;
  brands: Brand[];
}

export interface CreateCatalogItemBody {
  name: string;
  icon?: string | null;
}

export interface AddBrandBody {
  catalogItemId: number;
  name: string;
}

export interface AccountAsset {
  account: string;
  itemType: string;
  brand: string;
  count: number;
  lastDate?: string | null;
}

export type GetRequestsParams = {
  repUsername?: string;
};

export type GetEventsParams = {
  repUsername?: string;
};

export type GetTransfersParams = {
  repUsername?: string;
};

export type GetAccountAssetsParams = {
  repUsername: string;
};