# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BrewAsset POS & WMS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /requests:
    get:
      operationId: getRequests
      tags: [requests]
      summary: Get all requests
      parameters:
        - name: repUsername
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Request list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PosRequest"
    post:
      operationId: createRequest
      tags: [requests]
      summary: Submit a POS request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRequestBody"
      responses:
        "200":
          description: Request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PosRequest"
  /requests/{id}/fulfill:
    post:
      operationId: fulfillRequest
      tags: [requests]
      summary: Mark request as fulfilled
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Request fulfilled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /requests/{id}/cancel:
    post:
      operationId: cancelRequest
      tags: [requests]
      summary: Cancel a request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Request cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /events:
    get:
      operationId: getEvents
      tags: [events]
      summary: Get all events
      parameters:
        - name: repUsername
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Events list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create an event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventBody"
      responses:
        "200":
          description: Event created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{id}/approve:
    post:
      operationId: approveEvent
      tags: [events]
      summary: Approve an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /events/{id}/cancel:
    post:
      operationId: cancelEvent
      tags: [events]
      summary: Cancel an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /events/{id}/staff:
    post:
      operationId: toggleEventStaff
      tags: [events]
      summary: Assign or remove staff from event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToggleStaffBody"
      responses:
        "200":
          description: Staff updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /events/{id}/approve-pos:
    post:
      operationId: approvePOS
      tags: [events]
      summary: Approve POS materials for event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: POS approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /catalog/items:
    get:
      operationId: getCatalogItems
      tags: [catalog]
      summary: Get item types
      responses:
        "200":
          description: Item types
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CatalogItem"
    post:
      operationId: createCatalogItem
      tags: [catalog]
      summary: Add item type
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCatalogItemBody"
      responses:
        "200":
          description: Item type created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CatalogItem"
  /catalog/items/{id}:
    delete:
      operationId: deleteCatalogItem
      tags: [catalog]
      summary: Remove item type
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Item type removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /catalog/brands:
    post:
      operationId: addBrand
      tags: [catalog]
      summary: Add brand to item type
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddBrandBody"
      responses:
        "200":
          description: Brand added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /catalog/brands/{id}:
    delete:
      operationId: deleteBrand
      tags: [catalog]
      summary: Remove brand
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Brand removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
    LoginResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
      required: [user]
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
          enum: [rep, warehouse, marketing]
        displayName:
          type: string
      required: [id, username, role, displayName]
...
    PosRequest:
      type: object
      properties:
        id:
          type: integer
        repUsername:
          type: string
        customer:
          type: string
        itemType:
          type: string
        brand:
          type: string
        quantity:
          type: integer
        status:
          type: string
          enum: [Pending, Fulfilled, Cancelled]
        createdAt:
          type: string
        fulfilledAt:
          type: string
          nullable: true
      required: [id, repUsername, customer, itemType, brand, quantity, status, createdAt]
    CreateRequestBody:
      type: object
      properties:
        customer:
          type: string
        itemType:
          type: string
        brand:
          type: string
        quantity:
          type: integer
      required: [customer, itemType, brand, quantity]
    Event:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        account:
          type: string
        brand:
          type: string
          nullable: true
        date:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        notes:
          type: string
          nullable: true
        posItems:
          type: array
          items:
            type: string
        staffRequested:
          type: integer
        repUsername:
          type: string
        status:
          type: string
          enum: [pending, approved, cancelled]
        staffAssigned:
          type: array
          items:
            type: string
        posApproved:
          type: boolean
        posRequested:
          type: boolean
        createdAt:
          type: string
        approvedAt:
          type: string
          nullable: true
        approvedBy:
          type: string
          nullable: true
      required: [id, title, account, date, startTime, endTime, posItems, staffRequested, repUsername, status, staffAssigned, posApproved, posRequested, createdAt]
    CreateEventBody:
      type: object
      properties:
        title:
          type: string
        account:
          type: string
        brand:
          type: string
          nullable: true
        date:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        notes:
          type: string
          nullable: true
        posItems:
          type: array
          items:
            type: string
        staffRequested:
          type: integer
      required: [title, account, date, startTime, endTime, posItems, staffRequested]
    ToggleStaffBody:
      type: object
      properties:
        staffName:
          type: string
      required: [staffName]
...
    CatalogItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        icon:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
        isDefault:
          type: boolean
        brands:
          type: array
          items:
            $ref: "#/components/schemas/Brand"
      required: [id, name, isDefault, brands]
    Brand:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        catalogItemId:
          type: integer
        isDefault:
          type: boolean
      required: [id, name, catalogItemId, isDefault]
    CreateCatalogItemBody:
      type: object
      properties:
        name:
          type: string
        icon:
          type: string
          nullable: true
      required: [name]
    AddBrandBody:
      type: object
      properties:
        catalogItemId:
          type: integer
        name:
          type: string
      required: [catalogItemId, name]
...