# lib/api-zod/src/generated/types/createRequestBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BrewAsset POS & WMS API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateRequestBody {
  customer: string;
  itemType: string;
  brand: string;
  quantity: number;
}